requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.1
djangorestframework==3.14.0
Faker==24.1.0
Markdown==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
validate-docbr==1.10.0
